Implement IP 4 to 6 mapping on HostIP listen addresses
In some cases, the docker container can specify a HostIP listen address
to listen to. It might be desirable to allow converting containers
listening on IPv4 only addresses to IPv6 by using a network mapping.

The IPv4 address is mapped to IPv6 by matching it to an IPv4 CIDR
provided via a command-line flag. Multiple mapping is possible.